docs: move fruit utils for external use
