+++ apps/docs/src/utils/fruits.ts
@@ -1,0 +1,285 @@
+declare global {
+  interface Array<T> {
+    shuffled(count?: number): Array<T>
+  }
+}
+
+// eslint-disable-next-line
+Array.prototype.shuffled = function (count = 5) {
+  const result = this.sort(() => Math.random() - Math.random()).slice(0, count)
+  return count === 1 ? result[0] : result
+}
+
+export const fruits: Array<{
+  name: string
+  description: string
+  value: string
+  image: string
+}> = [
+  {
+    name: 'Ananas',
+    description: 'Tropisk frukt med taggigt skal',
+    value: 'ananas',
+    image:
+      'https://upload.wikimedia.org/wikipedia/commons/thumb/c/cb/Pineapple_and_cross_section.jpg/320px-Pineapple_and_cross_section.jpg',
+  },
+  {
+    name: 'Apelsin',
+    description: 'Citrusfrukt med orange skal',
+    value: 'apelsin',
+    image:
+      'https://upload.wikimedia.org/wikipedia/commons/thumb/c/c4/Orange-Fruit-Pieces.jpg/320px-Orange-Fruit-Pieces.jpg',
+  },
+  {
+    name: 'Aprikos',
+    description: 'Stenfrukt med orange färg',
+    value: 'aprikos',
+    image:
+      'https://upload.wikimedia.org/wikipedia/commons/thumb/2/2a/Apricot_and_cross_section.jpg/320px-Apricot_and_cross_section.jpg',
+  },
+  {
+    name: 'Avokado',
+    description: 'Grönsaksfrukt med krämig konsistens',
+    value: 'avokado',
+    image:
+      'https://upload.wikimedia.org/wikipedia/commons/thumb/2/2a/Avocado_IMGP1082.jpg/320px-Avocado_IMGP1082.jpg',
+  },
+  {
+    name: 'Banan',
+    description: 'Långsmal frukt med gult skal',
+    value: 'banan',
+    image:
+      'https://upload.wikimedia.org/wikipedia/commons/thumb/8/8a/Banana-Single.jpg/320px-Banana-Single.jpg',
+  },
+  {
+    name: 'Björnbär',
+    description: 'Små svarta bär',
+    value: 'bjornbar',
+    image:
+      'https://upload.wikimedia.org/wikipedia/commons/thumb/c/cc/Blackberries_%28Rubus_fruticosus%29.jpg/320px-Blackberries_%28Rubus_fruticosus%29.jpg',
+  },
+  {
+    name: 'Blåbär',
+    description: 'Små blå bär',
+    value: 'blabar',
+    image:
+      'https://upload.wikimedia.org/wikipedia/commons/1/15/Blueberries.jpg',
+  },
+  {
+    name: 'Carambola',
+    description: 'Stjärnformad exotisk frukt',
+    value: 'carambola',
+    image:
+      'https://upload.wikimedia.org/wikipedia/commons/thumb/6/62/Averrhoa_carambola_ARS_k5735-7.jpg/160px-Averrhoa_carambola_ARS_k5735-7.jpg',
+  },
+  {
+    name: 'Citron',
+    description: 'Citrusfrukt med gult skal',
+    value: 'citron',
+    image:
+      'https://upload.wikimedia.org/wikipedia/commons/thumb/e/e4/Lemon.jpg/320px-Lemon.jpg',
+  },
+  {
+    name: 'Clementin',
+    description: 'Liten citrusfrukt med löst skal',
+    value: 'clementin',
+    image:
+      'https://upload.wikimedia.org/wikipedia/commons/thumb/a/a9/Oh_my_darling.jpg/320px-Oh_my_darling.jpg',
+  },
+  {
+    name: 'Drakfrukt',
+    description: 'Exotisk frukt med rött skal',
+    value: 'drakfrukt',
+    image:
+      'https://upload.wikimedia.org/wikipedia/commons/thumb/4/43/Pitaya_cross_section_ed2.jpg/294px-Pitaya_cross_section_ed2.jpg',
+  },
+  {
+    name: 'Granatäpple',
+    description: 'Frukt med många kärnor',
+    value: 'granatapple',
+    image:
+      'https://upload.wikimedia.org/wikipedia/commons/thumb/f/fa/Pomegranate_fruit_-_whole_and_piece_with_arils.jpg/320px-Pomegranate_fruit_-_whole_and_piece_with_arils.jpg',
+  },
+  {
+    name: 'Grapefrukt',
+    description: 'Stor citrusfrukt med rosa eller gult kött',
+    value: 'grapefrukt',
+    image:
+      'https://upload.wikimedia.org/wikipedia/commons/thumb/d/d0/Citrus_paradisi_%28Grapefruit%2C_pink%29_white_bg.jpg/320px-Citrus_paradisi_%28Grapefruit%2C_pink%29_white_bg.jpg',
+  },
+  {
+    name: 'Hallon',
+    description: 'Röda bär som växer på buskar',
+    value: 'hallon',
+    image:
+      'https://upload.wikimedia.org/wikipedia/commons/thumb/2/2b/Raspberries_%28Rubus_idaeus%29.jpg/320px-Raspberries_%28Rubus_idaeus%29.jpg',
+  },
+  {
+    name: 'Jordgubbe',
+    description: 'Röda bär med frön på utsidan',
+    value: 'jordgubbe',
+    image:
+      'https://upload.wikimedia.org/wikipedia/commons/thumb/e/e1/Strawberries.jpg/320px-Strawberries.jpg',
+  },
+  {
+    name: 'Kiwi',
+    description: 'Frukt med hårig brun skal och grönt kött',
+    value: 'kiwi',
+    image:
+      'https://upload.wikimedia.org/wikipedia/commons/thumb/b/b8/Kiwi_%28Actinidia_chinensis%29_1_Luc_Viatour.jpg/320px-Kiwi_%28Actinidia_chinensis%29_1_Luc_Viatour.jpg',
+  },
+  {
+    name: 'Kokosnöt',
+    description: 'Stor nöt med hårt skal',
+    value: 'kokosnot',
+    image:
+      'https://upload.wikimedia.org/wikipedia/commons/thumb/f/f1/Coconuts_-_single_and_cracked_open.jpg/320px-Coconuts_-_single_and_cracked_open.jpg',
+  },
+  {
+    name: 'Körsbär',
+    description: 'Små röda stenfrukter',
+    value: 'korsbar',
+    image:
+      'https://upload.wikimedia.org/wikipedia/commons/thumb/b/bb/Cherry_Stella444.jpg/320px-Cherry_Stella444.jpg',
+  },
+  {
+    name: 'Lime',
+    description: 'Liten grön citrusfrukt',
+    value: 'lime',
+    image:
+      'https://upload.wikimedia.org/wikipedia/commons/thumb/d/d3/Citrus_×aurantiifolia927505341.jpg/320px-Citrus_×aurantiifolia927505341.jpg',
+  },
+  {
+    name: 'Litchi',
+    description: 'Liten frukt med tunt rosa skal',
+    value: 'litchi',
+    image:
+      'https://upload.wikimedia.org/wikipedia/commons/thumb/5/5c/Lychee_fruits_and_seed.jpg/320px-Lychee_fruits_and_seed.jpg',
+  },
+  {
+    name: 'Mandarin',
+    description: 'Liten orange citrusfrukt',
+    value: 'mandarin',
+    image:
+      'https://upload.wikimedia.org/wikipedia/commons/thumb/4/49/Mandarin_Oranges_%28Citrus_Reticulata%29.jpg/320px-Mandarin_Oranges_%28Citrus_Reticulata%29.jpg',
+  },
+  {
+    name: 'Mango',
+    description: 'Söt exotisk frukt med stor kärna',
+    value: 'mango',
+    image:
+      'https://upload.wikimedia.org/wikipedia/commons/thumb/4/49/Mango_-_single.jpg/320px-Mango_-_single.jpg',
+  },
+  {
+    name: 'Melon',
+    description: 'Stor frukt med saftigt kött',
+    value: 'melon',
+    image:
+      'https://upload.wikimedia.org/wikipedia/commons/thumb/f/ff/Muskmelon.jpg/320px-Muskmelon.jpg',
+  },
+  {
+    name: 'Nektarin',
+    description: 'Slät variant av persika',
+    value: 'nektarin',
+    image:
+      'https://upload.wikimedia.org/wikipedia/commons/thumb/9/9e/Autumn_Red_peaches.jpg/320px-Autumn_Red_peaches.jpg',
+  },
+  {
+    name: 'Papaya',
+    description: 'Exotisk frukt med orange kött',
+    value: 'papaya',
+    image:
+      'https://upload.wikimedia.org/wikipedia/commons/thumb/6/6b/Papaya_cross_section_BNC.jpg/320px-Papaya_cross_section_BNC.jpg',
+  },
+  {
+    name: 'Passionsfrukt',
+    description: 'Frukt med många kärnor och syrligt kött',
+    value: 'passionsfrukt',
+    image:
+      'https://upload.wikimedia.org/wikipedia/commons/thumb/2/21/Passion_fruits_-_whole_and_halved.jpg/320px-Passion_fruits_-_whole_and_halved.jpg',
+  },
+  {
+    name: 'Persika',
+    description: 'Mjuk stenfrukt med ludet skal',
+    value: 'persika',
+    image:
+      'https://upload.wikimedia.org/wikipedia/commons/thumb/9/9e/Autumn_Red_peaches.jpg/320px-Autumn_Red_peaches.jpg',
+  },
+  {
+    name: 'Physalis',
+    description: 'Liten frukt som växer i pappershölje',
+    value: 'physalis',
+    image:
+      'https://upload.wikimedia.org/wikipedia/commons/thumb/7/7b/Physalis_peruviana_calix_open_close-up.jpg/300px-Physalis_peruviana_calix_open_close-up.jpg',
+  },
+  {
+    name: 'Plommon',
+    description: 'Söt eller syrlig stenfrukt',
+    value: 'plommon',
+    image:
+      'https://upload.wikimedia.org/wikipedia/commons/thumb/7/7d/Bluebyrd_plum.jpg/167px-Bluebyrd_plum.jpg',
+  },
+  {
+    name: 'Päron',
+    description: 'Avlång frukt med smal midja',
+    value: 'paron',
+    image:
+      'https://upload.wikimedia.org/wikipedia/commons/thumb/c/cf/Pears.jpg/393px-Pears.jpg',
+  },
+  {
+    name: 'Rambutan',
+    description: 'Exotisk frukt med hårig skal',
+    value: 'rambutan',
+    image:
+      'https://upload.wikimedia.org/wikipedia/commons/a/ae/Rambutan_Fruit.jpg',
+  },
+  {
+    name: 'Röda vinbär',
+    description: 'Små röda bär i klasar',
+    value: 'roda vinbar',
+    image:
+      'https://upload.wikimedia.org/wikipedia/commons/thumb/2/2e/Hjulsta_koloni_2010h.jpg/256px-Hjulsta_koloni_2010h.jpg',
+  },
+  {
+    name: 'Sharon',
+    description: 'Persikoliknande frukt med fast kött',
+    value: 'sharon',
+    image:
+      'https://upload.wikimedia.org/wikipedia/commons/9/95/Diospyros_kaki_-_persimmon_at_Paro_during_LGFC_-_Bhutan_2019_%283%29.jpg',
+  },
+  {
+    name: 'Stjärnfrukt',
+    description: 'Stjärnformad frukt med syrligt kött',
+    value: 'stjarnfrukt',
+    image:
+      'https://upload.wikimedia.org/wikipedia/commons/thumb/3/3a/Carambola_Starfruit.jpg/320px-Carambola_Starfruit.jpg',
+  },
+  {
+    name: 'Svarta vinbär',
+    description: 'Små svarta bär i klasar',
+    value: 'svarta vinbar',
+    image:
+      'https://upload.wikimedia.org/wikipedia/commons/1/17/Blackcurrants2.jpg',
+  },
+  {
+    name: 'Vattenmelon',
+    description: 'Stor frukt med rött, saftigt kött',
+    value: 'vattenmelon',
+    image:
+      'https://upload.wikimedia.org/wikipedia/commons/thumb/e/e4/Water_melon_2015.jpg/320px-Water_melon_2015.jpg',
+  },
+  {
+    name: 'Vindruvor',
+    description: 'Små gröna eller blå frukter i klasar',
+    value: 'vindruvor',
+    image:
+      'https://upload.wikimedia.org/wikipedia/commons/thumb/8/85/Thompson_seedless_grapes.JPG/320px-Thompson_seedless_grapes.JPG',
+  },
+  {
+    name: 'Äpple',
+    description: 'Rund frukt med kärnhus',
+    value: 'apple',
+    image:
+      'https://upload.wikimedia.org/wikipedia/commons/thumb/1/15/Red_Apple.jpg/320px-Red_Apple.jpg',
+  },
+]
